Simpler entry of assignments and reports
Simplified entry of assignments and reports through the web interface.

On the report form, begintime and endtime has been broken up into one date field and two time entries.
Jobs are populated with available jobs both when editing reports and when editing jobs.

Totalduration is added to make our data more Views-friendly. This makes it possible to list the total duration
in a view without resorting to the views_calc module.

SimpleTest tests that verify posting a job, a simple report, a rapport past midnight and reports on the spring and autumn DST shift.

Change-Id: Ie0045628fc6694067052f11ce1e00dbc95aab0e6