143: Fixes first time sync issue.
Adds tests and fix to ensure that we sync at the first login. The
cause of failure was that curl writes its cookiejar when the curl
handle is closed and we did not close the bot and curl handle after
first login and before first sync.

TZIntellitimeBot's constructor now throws an exception when the
cookie-file is empty or missing.

Change-Id: Icd3b041bb76a4b9d7785f41e59c2bc2bf1b9279c