172: Handle timezones in drupal
Uses the date_api and date_timezone modules to make sure we handle timezones
correctly everywhere in the code. New dates are created using the site timezone
and the wrapper function tzbase_make_date. The dates are standard php DateTime
objects and can therefor easily be formatted as needed.

Done
- Excelexport
- tzbase
- tzintellitime
- tzintellitime_sync
- tzreports
- updated translations

Change-Id: If998a3e5301a3d96bdd0e310e1df9eff8c555bdf