Fixes recently introduced time field parsing bug.
In 6be0d4539b84a35e8c799c642fe81dfb9ecc88ce, we mistakenly changed
'if(!$td)' to 'if(!empty($td))' instead of the correct
'if(empty($td))' in the private method parse_time_field of
TZIntellitimeParser.

This commit fixes the comparison and adds tests for
TZIntellitimeParser::parse_reports() to make sure that it does not
happen again.

Change-Id: Icf1a1bf9e7190ad3c9affe4fc7a20d18cf7a9f87