Fixes totalduration bug in tzproto and tzbase.
In tzproto, a report that spans over midnight gets its endtime
incorrectly set to less than its begintime, thus ending before it
starts. This causes its totalduration field to be calculated to 0.

This commit fixes the original bug, adds extra watchdogs to tzbase to
allow us to detect this in the future, and adds an update of tzbase
that corrects the relevant time report entries in the db.

Change-Id: I9e5f012f9dc4550d818ad1ea74762b7fbf955e73
Reviewed-on: http://code.timezynk.com/gerrit/625
Reviewed-by: Johan Heander <johan@timezynk.com>
Tested-by: Johan Heander <johan@timezynk.com>