Refactors loggers to use log severity levels.
This commit introduces log severity levels (as per RFC 3164) ranging
from DEBUG to EMERGENCY. It refactors the filtering on specific
exceptions used in the SMS logger into filtering on specific log
levels, which should be less prone to change.

This also sets the stage for moving the logging facilities to tzbase
or a separate module, with appropriate extensions in the relevant
modules.

Change-Id: I178c0ffff033a6de8870bf921dcbc857a8763f37
Reviewed-on: http://code.timezynk.com/gerrit/697
Reviewed-by: Mikael Ohlson <mikael.ohlson@timezynk.com>
Tested-by: Mikael Ohlson <mikael.ohlson@timezynk.com>