Removes the possibility to add new intellitime users on login.
As a feature of the intellitime module, it was possible to create a
new user if the given user-name and password could be authenticated by
the intellitime server. However, this caused a lot of problems, mostly
where administration had added the wrong user-name and/or password and
the user instead logged in with the correct info or an old user-name
and password. Not that big a problem in it self, but that way the
manager and mobile phone number of that user was unset, making it hard
to contact and track them.

This commit disables that feature.

Change-Id: Iea153188014ca88239fc42e496c0d7bede981713
Reviewed-on: http://code.timezynk.com/gerrit/730
Tested-by: Johan Heander <johan@timezynk.com>
Reviewed-by: Johan Heander <johan@timezynk.com>